# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Backend API",
    "description": "This API exposes endpoints to manage demo.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "Bearer_Authentication": []
    }
  ],
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/packages/{id}": {
      "get": {
        "tags": [
          "package-controller"
        ],
        "operationId": "getPackageById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PackageEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobroles/popular": {
      "get": {
        "tags": [
          "job-role-controller"
        ],
        "operationId": "getPopularJobRole",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseJobRoleEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationEntity": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "job": {
            "$ref": "#/components/schemas/JobEntity"
          },
          "freelancer": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "cvUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ]
          }
        }
      },
      "CVEntity": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cvUrl": {
            "type": "string"
          },
          "freelancer": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "EducationEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "type": "string"
          }
        }
      },
      "IndustryEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "JobEntity": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobTitle": {
            "type": "string"
          },
          "jobDescription": {
            "type": "string"
          },
          "requirementDescription": {
            "type": "string"
          },
          "desirableDescription": {
            "type": "string"
          },
          "benefitDescription": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "jobBenefits": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "jobRole": {
            "$ref": "#/components/schemas/JobRoleEntity"
          },
          "jobTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobTypeEntity"
            }
          },
          "minimumSalary": {
            "type": "number",
            "format": "double"
          },
          "maximumSalary": {
            "type": "number",
            "format": "double"
          },
          "education": {
            "$ref": "#/components/schemas/EducationEntity"
          },
          "experience": {
            "type": "string"
          },
          "vacancies": {
            "type": "integer",
            "format": "int32"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "jobLevel": {
            "$ref": "#/components/schemas/JobLevelEntity"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "logoPath": {
            "type": "string"
          },
          "employer": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationEntity"
            }
          },
          "favoritedByUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserEntity"
            }
          }
        }
      },
      "JobLevelEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "type": "string"
          }
        }
      },
      "JobRoleEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "industry": {
            "$ref": "#/components/schemas/IndustryEntity"
          },
          "jobLevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobLevelEntity"
            }
          }
        }
      },
      "JobTypeEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobEntity"
            }
          }
        }
      },
      "SocialLinkEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "platform": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "EMPLOYER",
              "CANDIDATE"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "isSetup": {
            "type": "boolean"
          },
          "canPost": {
            "type": "boolean"
          },
          "socialProvider": {
            "type": "string",
            "enum": [
              "DATABASE",
              "FACEBOOK",
              "GOOGLE"
            ]
          },
          "providerId": {
            "type": "string"
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLinkEntity"
            }
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobEntity"
            }
          },
          "favoriteJobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobEntity"
            }
          },
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationEntity"
            }
          },
          "cvs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CVEntity"
            }
          },
          "purchasedPackages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PackageEntity": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "durationInDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PagedResponseJobRoleEntity": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobRoleEntity"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer_Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}